requirements.txt
alembic==1.11.1
anyio==3.7.0
boto3==1.26.162
botocore==1.29.162
certifi==2023.7.22
click==8.1.3
colorama==0.4.6
dnspython==2.3.0
email-validator==2.0.0.post2
exceptiongroup==1.1.1
fastapi~=0.115.0
greenlet
h11==0.14.0
httpcore==0.17.2
httptools
httpx==0.24.1
idna==3.7
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
Mako==1.2.4
MarkupSafe==2.1.3
orjson==3.9.15
postgres==4.0
psycopg2==2.9.6
psycopg2-binary==2.9.6
psycopg2-pool==1.1
pydantic~=2.9.2
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.7
PyYAML==6.0
s3transfer==0.6.1
six==1.16.0
sniffio==1.3.0
SQLAlchemy~=2.0.35
sqlalchemy2-stubs==0.0.2a34
sqlmodel~=0.0.22
starlette==0.36.2
typing_extensions==4.6.3
ujson==5.8.0
urllib3==2.0.7
uvicorn==0.22.0
watchfiles==0.19.0
websockets==11.0.3
win32-setctime==1.1.0
python-multipart
pytz~=2024.2
jwt~=1.3.1
pydantic-settings~=2.5.2
moviepy~=1.0.3
requests~=2.32.3
line-bot-sdk~=3.13.0
PyJWT~=2.9.0